// backend/model/dashboard/fileUpload_mdl.ts
// export type PostFile = {
//     file_id: string;
//     file_seq: string;
//     file_name: string;
//     file_size: string;
//     file_path: string;
//     sevr_file_name: string;
//     use_yn: string;
//     insert_id: string;
//     insert_date: string;
//     update_id: string;
//     update_date: string;
//   };

  export type PostFile = {
    file_id: string;
    file_seq: string;
    file_name: string;
    file_size: string;
    file_path: string;
    sevr_file_name: string;
    use_yn: string;
  };

  export type GetFile = {
    file_id: string;
    file_seq: string;
    file_name: string;
    file_size: string;
    file_path: string;
    sevr_file_name: string;
    use_yn: string;
  };